+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Fix bug when adding smooth line (missing import for `geom_smooth`).
 * Support for themes and scales (color/fill) from [{hrbrthemes}](https://github.com/hrbrmstr/hrbrthemes).
+* Set labels for fill, color and size aesthetics.
 
 
 # esquisse 0.2.0